net: avoid closing TCP session because of any outcoming RST
Current u-boot TCP support is really limited and can handle
only 1 client at 1 moment because whole TCP state is stored
globally for all incoming connections. In case we're already
handling 1 connection and, for example, receiving SYN for another
one, TCP stack is answering RST, but after my change here:
http://patchwork.ozlabs.org/project/uboot/patch/20230412184931.1422789-1-dimorinny@google.com/,
it started also resetting the TCP state which is wrong because
now any incoming connection attempt is breaking the current TCP
session.

The final solution is going to be splitting existing global TCP
state into the instances on a per-connection basis, so we can
handle TCP state for different clients separately.

For now fixing the particular problem to allow filter out traffic
properly with the current limitations.

Will be upstreamed as a part of: b/295225295

Bug: 294901799
Test: adb connect to fastboot during fastboot erase is ok
Change-Id: I0897915e5300a5b2571220e8024ad6e1fba49482
Signed-off-by: Dmitrii Merkurev <dimorinny@google.com>