net: handle hanging TCP connections by introducing timeout
Existing u-boot TCP stack has global state for all connections,
so it can handle only 1 connection at time. But if someone
connected and disappeared we cannot accept any new connections.

The proper fix for this problem is to split TCP state on
per-connection basis, so we will be able to handle multiple
connections at the same time, but this is a big change which
is not going to be useful in a long term since existing network
stack is going to be replaced with lwip soon. Related stack:

https://patchwork.ozlabs.org/project/uboot/list/?series=374882&state=*

Fixing it by introducing 5 seconds timeout, so now TCP connection
can be interrupted if it does not interact with TCP app for 5
seconds or more

Test: 1000 runs of FastbootRebootTest on abtd with 0 failures
Test: 100 runs of FastbootFlashingTest on abtd with 0 failures
Test: flashed / rebooted cvd 100 times locally
Bug: 303236495
Change-Id: I88e22eb465dabc52379f4524590f4d4acf867902
Signed-off-by: Dmitrii Merkurev <dimorinny@google.com>