virtio-console: don't interact with the control queue in single-port mode
Also cleans up some error returns that were not being handled.

This fixes the single-port flow with QEMU which otherwise would abort on
interaction with the control queue. I'm still not sure how that ever
worked under crosvm or x86_64 qemu, since the control queue is NULL in
single-port mode.

The multiport flow is still broken on aarch64 qemu.

Bug: b/351052783
Change-Id: I1cf19e9830a929c866ff6866a862335bd39f77d1
Test: tools/bazel run //u-boot:qemu_aarch64_compile_commands && tools/bazel run //u-boot:qemu_aarch64_dist
Test: Run aarch64 under qemu on Cuttlefish