Implement UEFI protocol for GBL AB slot metadata
Create an implementation for the custom GBL AB slot protocol that
stores slot metadata in virtio:0#misc at block 1.

Note: this implementation is intended for Cuttlefish targets only for
integration testing. It is NOT intended for actual, production use.

Includes an addition to bootefi selftest to verify behavior.

Tests: bootefi selftest passes checks

Bug: b/348426365

Change-Id: I5eade93bbd9731b31791fc753992fa7654100f87