board: hvm: add HVM board to U-Boot
Intdroducing new HVM board that is meant to be started from Xen PVH
domain.
The main approach is that U-Boot provides binary file which should be
the part of the PVH kernel which acts as a contaier for Xen PVH
domain. For example - mini-os.
This kernel task is to boot by Xen as PVH domain, provide
initialization of the tables, stach and memory and then jump to U-Boot
entry point.

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>